requirements.txt
# local package
# -e .

# external requirements
numpy~=1.19.5,<1.21.0
pandas~=1.1.5
joblib>=1.0.0
typing-extensions>=3.10.0
pylint~=2.11.1
matplotlib~=3.3.4
tox~=3.24.4
pytest~=6.2.5
pathlib~=1.0.1
pydantic~=1.8.2
strictyaml~=1.5.0
keras~=2.6.0
scikit-learn==0.21.2
scipy~=1.3.1
mlflow~=1.21.0
mypy~=0.910
flake8==4.0.1
flake8-junit-report==2.1.0
# tensorflow~=2.6.1

# {% if cookiecutter.python_interpreter != 'python3' %}
# # backwards compatibility
# pathlib2
# {% endif %}
